viewer: read run listings from the derived store
/api/runs, /api/results, /api/ladder, /api/models, /api/fleet, /api/info
and the run page read rows from ClickHouse instead of opening every run
directory. The seam is one module, `runner/viewer/clickhouse.ts`, and it
is deliberately at "fetch me these rows" rather than "answer this
question": a store that answered questions would be a second
implementation of every derivation the viewer already has, free to
disagree with the first about a published number. Only one aggregate is
pushed into SQL — the latest reading per run — and it is stated twice,
in SQL and in TypeScript, with a test that says what it must mean.

Response shapes are unchanged. `runner/test/viewer-clickhouse.test.ts`
pins a RunRow built from stored rows against the one `readRun` builds
from the files, field for field in the same order, and does the same for
the state series and the movement intentions. Verified further against a
live ClickHouse over two real runs: /api/runs, /api/results,
/api/models, /api/info and /api/run/:id come back byte-identical from
the ClickHouse store and the in-memory one.

Two defects that comparison found. `trajectory_mtime` was Int64, which
truncated the sub-millisecond precision a stat carries and the viewer
puts on the wire; it is Float64. And "the latest sample" broke a
same-millisecond tie by whatever a scan reached first; the key is
(ts, seq) on both sides now.

Deleted, because nothing else used them: `totalsCache`, `runReadCache`,
`fact-store.ts` and `readRunFactsCached` with the caching half of
`runs.ts`, and the two test files that covered them. What STAYS on
files, and must, is everything that asks about a process writing right
now: `listRuns`/`readRun`/`readMoves` for the fleet supervisor
(`archive.ts`), the live map (`positions.ts`) and the account ledger on
/api/fleet — each of which prefilters to the runs touched inside its own
window rather than scanning the tree — plus the per-run trajectory tail
behind /entries and the SSE stream. Retiring the per-run sqlite is a
later item.

With no CLICKHOUSE_URL the viewer builds the same rows in memory by
running the collector over the runs directory. Not a mock: it is the
same ingestion, so a row means the same thing, and it is what a bare
clone and every test get.

+++ collector/schema.sql
@@ -4,10 +4,11 @@
 --
 -- Conventions that hold for every table below, and why:
 --
---   * Timestamps are `Int64` epoch milliseconds, not `DateTime64`. That is
---     what the runner writes and what every API body already carries, so a
---     round trip through this store changes no byte a client sees. Read them
---     with `fromUnixTimestamp64Milli(ts)` when a human wants a date.
+--   * Timestamps are epoch milliseconds — `Int64`, or `Float64` where the
+--     source carries sub-millisecond precision — not `DateTime64`. That is what
+--     the runner writes and what every API body already carries, so a round trip
+--     through this store changes no byte a client sees. Read them with
+--     `fromUnixTimestamp64Milli(ts)` when a human wants a date.
 --   * No `PARTITION BY`. ReplacingMergeTree deduplicates only within a
 --     partition, and any time-based partition would split a run that crosses
 --     the boundary and keep both copies of its rows forever. The corpus is
@@ -65,7 +66,10 @@ CREATE TABLE IF NOT EXISTS wrathbench.runs
   -- The trajectory file's own size and mtime at ingestion. Liveness is decided
   -- from the mtime against `now` exactly as it is from the file today.
   trajectory_bytes      Int64,
-  trajectory_mtime      Int64,
+  -- Float64, not Int64: a stat's `mtimeMs` carries sub-millisecond precision
+  -- and the viewer puts it on the wire verbatim, so truncating it here would
+  -- change a byte of every run listing.
+  trajectory_mtime      Float64,
   ingested_at           Int64
 )
 ENGINE = ReplacingMergeTree(ingested_at)
